Добавленный во фронте станица регистрации и входа
