+++ backend/src/main/java/ru/sber/entity/ERegularity.java
