+++ backend/src/main/java/ru/sber/entity/EStatus.java
