+++ backend/src/main/java/ru/sber/entity/Task.java
@@ -1,0 +1,55 @@
+package ru.sber.entity;
+
+import java.time.LocalDateTime;
+import com.fasterxml.jackson.annotation.JsonFormat;
+import jakarta.persistence.*;
+import lombok.NoArgsConstructor;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+@NoArgsConstructor
+@Entity
+@Table (name="Tasks")
+public class Task {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.AUTO)
+    private Long id;
+
+    @Column(nullable = false)
+    private String title;
+
+    @Column(nullable = false)
+    private String description;
+
+    @Column(nullable = false)
+    @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
+    private LocalDateTime date;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private EStatus status;
+
+    @Column(nullable = false)
+    @Enumerated(EnumType.STRING)
+    private EPriority priority;
+
+    @Column(nullable = false)
+    private boolean archive;
+
+   @Column(nullable = false)
+   @Enumerated(EnumType.STRING)
+    private ERegularity regularity;
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "id_category", nullable = false)
+    private Category category;
+
+
+    @ManyToOne(fetch = FetchType.LAZY)
+    @JoinColumn(name = "id_user", nullable = false)
+    private User user;
+
+}
