--- src/main/java/ru/sber/entity/Category.java
@@ -1,31 +1,0 @@
-package ru.sber.entity;
-
-
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import jakarta.annotation.Nonnull;
-import jakarta.persistence.*;
-
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Table(name = "Categories",
-        uniqueConstraints = {
-                @UniqueConstraint(columnNames = "name")
-        })
-public class Category {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @JsonIgnoreProperties
-    private Long id;
-
-    @Column(nullable = false)
-    private String name;
-
-}
